requirements.txt
# Core AI/ML
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0

# Vector Store
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0

# Machine Learning
scikit-learn>=1.3.0

# Knowledge Graph
networkx>=3.3
pyvis>=0.3.2

# Research APIs
arxiv>=2.1.0
httpx>=0.27.0
requests>=2.32.0

# UI
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
numpy>=1.26.0,<2.0.0
pandas>=2.2.0
aiohttp>=3.9.0